[docs] 회원탈퇴 호출주소 병합
- 일반 로그인 회원가 카카오 로그인 회원의 회원탈퇴 API 호출주소를 하나로 병합
