+++ src/main/java/com/momo/common/exception/ErrorCode.java
@@ -11,7 +11,8 @@ public enum ErrorCode {
   INVALID_REQUEST("잘못된 요청입니다.", 400),
   INVALID_TOKEN("유효하지 않은 토큰입니다.", 400),
   EMAIL_SEND_FAILED("메일 발송을 실패했습니다.",400),
-  PROFILE_NOT_FOUND("프로필을 찾을 수 없습니다.",400);
+  PROFILE_NOT_FOUND("프로필을 찾을 수 없습니다.",400),
+  KAKAO_UNLINK_FAILED("카카오 계정 연동해제를 실패했습니다.",400);
 
   private final String message;
   private final int status;
